Refactor useBookmarks
