+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useTextField } from "./useTextField";
+export { useBookmarks } from "./useBookmarks";
